# docs/openapi.yaml
openapi: 3.1.0
info:
  title: todo
  description: ''
  version: 1.0.0
...
paths:
...
  /projects:
    post:
      summary: create
      deprecated: false
      description: ''
      tags: []
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type:
                    - string
                    - 'null'
              required:
                - title
            example:
              title: test todo
              description: test description
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID
                  title:
                    type: string
                  description:
                    type: string
                  createdDt:
                    type: string
                    description: UTC TIME
                    format: date-time
                  updatedDt:
                    type: string
                    description: UTC TIME
                    format: date-time
                required:
                  - id
                  - title
                  - description
                  - createdDt
                  - updatedDt
              example:
                id: 1
                title: test todo
                description: test description
                createdDt: '2026-03-09T07:40:41.580378'
                updatedDt: '2026-03-09T07:40:41.580389'
          headers: {}
        '422':
          $ref: '#/components/responses/422'
          description: ''
      security: []
  /projects/{project_id}/todos:
    post:
      summary: create todo
      deprecated: false
      description: ''
      tags: []
      parameters:
        - name: project_id
          in: path
          description: ''
          required: true
          example: '1'
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type:
                    - string
                    - 'null'
                priority:
                  type: string
                  default: low
                  enum:
                    - low
                    - medium
                    - high
              required:
                - title
                - priority
            example:
              title: test todo
              description: test description
              priority: medium
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  project_id:
                    type:
                      - integer
                      - 'null'
                    description: 속한 프로젝트 아이디
                  title:
                    type: string
                  description:
                    type:
                      - string
                      - 'null'
                  status:
                    type: boolean
                    description: 상태
                  priority:
                    type: string
                    default: low
                    enum:
                      - low
                      - medium
                      - high
                  deletedDt:
                    type:
                      - string
                      - 'null'
                    description: UTC TIME
                    format: date-time
                  createdDt:
                    type: string
                    description: UTC TIME
                    format: date-time
                  updatedDt:
                    type: string
                    description: UTC TIME
                    format: date-time
                required:
                  - id
                  - title
                  - status
                  - priority
                  - createdDt
                  - updatedDt
              example:
                id: 1
                projectId: 1
                title: test todo
                description: test description
                status: false
                priority: medium
                deletedDt: null
                createdDt: '2026-03-09T07:41:02.558989'
                updatedDt: '2026-03-09T07:41:02.558998'
          headers: {}
        '422':
          $ref: '#/components/responses/422'
          description: ''
      security: []
...
components:
...
  responses:
    '422':
      description: ''
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type:
                        - string
                        - 'null'
                    loc:
                      type: 'null'
                    msg:
                      type: string
                    input:
                      type:
                        - string
                        - 'null'
                  required:
                    - type
                    - loc
                    - msg
                    - input
              type:
                type: string
            required:
              - detail
              - type
          examples:
            예시 1:
              summary: 예시 1
              value:
                detail:
                  - type: missing
                    loc:
                      - body
                    msg: Field required
                    input: null
  securitySchemes: {}
servers: []
security: []